--- MagneticElastica/__init__.py
@@ -1,1 +1,0 @@
-from MagneticElastica.magnetic_forces import *
