+++ elastica/magnetic_forces.py
