+++ examples/MagneticElastica/magnetic_beam_analytical_solution.py
