+++ examples/MagneticElastica/post_processing.py
